+++ AzureKeyVault/AzureClient.cs
@@ -79,7 +79,7 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
                 else
                 {
                     logger.LogTrace("They are using a service principal to authenticate, generating the credentials");
-                    cred = new ClientSecretCredential(VaultProperties.TenantId, VaultProperties.ClientId, VaultProperties.ClientSecret);
+                    cred = new ClientSecretCredential(VaultProperties.TenantId, VaultProperties.ClientId, VaultProperties.ClientSecret, new ClientSecretCredentialOptions() { AuthorityHost = AzureCloudEndpoint });
                     logger.LogTrace("generated credentials", cred);
                 }
 
@@ -151,21 +151,42 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
         {
             try
             {
-                logger.LogInformation("Begin create vault...");
+                logger.LogInformation($"Begin create vault in Subscription {VaultProperties.SubscriptionId} with storepath = {VaultProperties.StorePath}");
 
-                SubscriptionResource subscription = await KvManagementClient.GetDefaultSubscriptionAsync();
-                ResourceGroupCollection resourceGroups = subscription.GetResourceGroups();
-                ResourceGroupResource resourceGroup = await resourceGroups.GetAsync(this.VaultProperties.ResourceGroupName);
+                logger.LogTrace($"getting subscription info for provided subscription id {VaultProperties.SubscriptionId}");
 
-                var vaults = resourceGroup.GetKeyVaults();
+                SubscriptionResource subscription = KvManagementClient.GetSubscriptionResource(SubscriptionResource.CreateResourceIdentifier(VaultProperties.SubscriptionId));
+                //ResourceGroupCollection resourceGroup = subscription.GetResourceGroup(VaultProperties.ResourceGroupName)
+                ResourceGroupResource resourceGroup = subscription.GetResourceGroup(VaultProperties.ResourceGroupName);
 
-                var loc = new AzureLocation(VaultProperties.VaultRegion);
+                AzureLocation loc;
+
+                var vaults = resourceGroup.GetKeyVaults();                
+                if (string.IsNullOrEmpty(VaultProperties.VaultRegion))
+                {
+                    try
+                    {
+                        logger.LogTrace($"no Vault Region location specified for new Vault, Getting available regions for resource group {resourceGroup.Data.Name}.");
+                        var locOptions = await resourceGroup.GetAvailableLocationsAsync();
+                        logger.LogTrace($"got location options for subscription {subscription.Data.SubscriptionId}", locOptions);
+                        loc = locOptions.Value.FirstOrDefault();
+                    }
+                    catch (Exception ex)
+                    {
+                        logger.LogError($"error retrieving default Azure Location: {ex.Message}", ex);
+                        throw;
+                    }
+                }
+                else
+                {
+                    loc = new AzureLocation(VaultProperties.VaultRegion);
+                }
 
                 var skuType = VaultProperties.PremiumSKU ? KeyVaultSkuName.Premium : KeyVaultSkuName.Standard;
 
                 var content = new KeyVaultCreateOrUpdateContent(loc, new KeyVaultProperties(new Guid(VaultProperties.TenantId), new KeyVaultSku(KeyVaultSkuFamily.A, skuType)));
-
-                var newVault = await vaults.CreateOrUpdateAsync(WaitUntil.Completed, VaultProperties.VaultName, content); //this takes a bit of time to run
+                
+                var newVault = await vaults.CreateOrUpdateAsync(WaitUntil.Completed, VaultProperties.VaultName, content);
 
                 return newVault.Value;
             }
@@ -267,7 +288,15 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
                         logger.LogTrace($"searching for vaults in subscription with ID {sub.Data.SubscriptionId}");
                         var vaults = sub.GetKeyVaults();
                         logger.LogTrace($"found {vaults.Count()} vaults.");
-                        vaultNames.AddRange(vaults.Select(v => v.Id.ToString()));
+
+                        foreach (var vault in vaults) {
+                            var splitId = vault.Id.ToString().Split("/", StringSplitOptions.RemoveEmptyEntries);
+                            // example resource identifier: /subscriptions/b3114ff1-bb92-45b6-9bd6-e4a1eed8c91e/resourceGroups/azure_sentinel_evaluation/providers/Microsoft.KeyVault/vaults/jv2-vault
+                            var subId = splitId[1];
+                            var resourceGroupName = splitId[3];
+                            var vaultName = splitId.Last();
+                            vaultNames.Add($"{subId}:{resourceGroupName}:{vaultName}");
+                        }                        
                     }
                 });